Revert "둘러보기 상세 급하게 추가"
