+++ Toonie/Common/Network/Service/Service.swift
@@ -1,7 +1,6 @@
 import Foundation
 
 struct TokenService : Requestable {
-    
     typealias NetworkData = Token
     static let shareInstance = TokenService()
